src/requirements.txt
ansible==2.2.1.0
appdirs==1.4.0
astroid==1.4.9
awscli==1.11.51
backports.functools-lru-cache==1.3
boto==2.46.1
boto3==1.4.4
botocore==1.5.14
cffi==1.9.1
colorama==0.3.7
configparser==3.5.0
cryptography==1.7.2
Django==1.10.5
django-cors-headers==2.0.2
djangorestframework==3.5.4
docutils==0.13.1
enum34==1.1.6
futures==3.0.5
idna==2.2
ipaddress==1.0.18
isort==4.2.5
Jinja2==2.8.1
jmespath==0.9.1
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mccabe==0.6.1
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.2
pycparser==2.17
pycrypto==2.6.1
pylint==1.6.5
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
wrapt==1.10.8
